+++ src/senaite/astm/tests/test_dca_vantage.py
@@ -7,7 +7,7 @@ from senaite.astm import codec
 from senaite.astm.constants import ACK
 from senaite.astm.constants import ENQ
 from senaite.astm.instruments import dca_vantage
-from senaite.astm.protocol import ASTMProtocol
+from senaite.astm.transports.astm.protocol import ASTMProtocol
 from senaite.astm.tests.base import ASTMTestBase
 from senaite.astm.wrapper import Wrapper