+++ src/senaite/astm/tests/test_genexpert.py
@@ -4,7 +4,7 @@ from senaite.astm import codec
 from senaite.astm.constants import ACK
 from senaite.astm.constants import ENQ
 from senaite.astm.instruments import genexpert
-from senaite.astm.protocol import ASTMProtocol
+from senaite.astm.transports.astm.protocol import ASTMProtocol
 from senaite.astm.tests.base import ASTMTestBase
 from senaite.astm.wrapper import Wrapper
 from unittest.mock import MagicMock